Setup daily job with  library
