+++ requirements.txt
@@ -12,6 +12,7 @@ beautifulsoup4~=4.11.2
 # sqlalchemy~=2.0.4
 # fastapi~=0.89.1
 # uvicorn~=0.20.0
+schedule~=1.1.0
 
 # Env
 python-dotenv~=1.0.0
